pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.16",
    "chromadb>=1.0.3",
    "email-validator>=2.2.0",
    "fastapi>=0.115.9",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "google-api-python-client>=2.166.0",
    "google-auth>=2.38.0",
    "google-auth-oauthlib>=1.2.1",
    "gunicorn>=23.0.0",
    "openai>=1.72.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "sqlite-utils>=3.38",
    "uvicorn>=0.34.0",
    "websockets>=15.0.1",
]
